Make warrior aware of level so that it can load the player.py. First tower is playable now
Needed to make warrior aware of level so that we can import the
player.py that the user provides. In rubywarrior, the player file could
be loaded elsewhere and the warrior could use it.  In python, I think we
need to import the player file in the warrior file where it's going to
be used